// ClientApp/src/models/RegisterInstitutionEntity.ts
export default class RegisterInstitutionEntity {
    institutionName: string = "";
    taxIdentificationNumber: string = "";
    country: string = "";
    city: string = "";
    streetAddress = "";
    streetNumber = "";
    postalCode = "";
    firstName = "";
    lastName = "";
    phoneNumber = "";
    email = "";
    password = "";
}